5.base-human/multi-Fusion/requirements.txt
opencv-python== 4.10.0.84
mediapipe==0.10.10
websocket-client==1.6.1
websockets==11.0.3
imutils==0.5.4
PyAudio==0.2.14
onnxruntime==1.19.2
